% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriquepresa-URL-functions.R
\name{dfGO}
\alias{dfGO}
\title{generation of a data frame with URLs from GO
URLs for a set of GO terms}
\usage{
dfGO(df)
}
\arguments{
\item{df}{data frame that is the result of some GO enrichment analysis and with GO terms as rownames}
}
\value{
A data frame with URLs from GO for every GO term
}
\description{
This function receives a data frame that is the result of some GO enrichment analysis and makes a
  data frame with URLs from GO for every GO term. The GO terms must be as rownames.
}
\seealso{
Other URL generation: 
\code{\link{ensemblurl}()},
\code{\link{entrezurl}()}
}
\concept{URL generation}
