% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriquepresa-data.R
\docType{data}
\name{edgeR_df}
\alias{edgeR_df}
\title{edgeR_df}
\format{
data.frame
}
\source{
\url{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA517180}
}
\description{
Data frame with the result of doing a differential expression analysis with edgeR package and the
   SummarizedExperiment PRJNA517180. This data frame is included as data in order to be able to do the
   overrepresentation analysis. 
In this data frame, the WORMBASE identifiers as rownames, gene symbols, logFC, logCPM, P-value,
   Benjamini-Hochberg adjusted P-value, and ENTREZ and ENSEMBL URLs are included.
}
\examples{
data(edgeR_df,package="enriquepresa")
}
\keyword{datasets}
