% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriquepresa-URL-functions.R
\name{ensemblurl}
\alias{ensemblurl}
\title{ENSEMBL URL generation
URL for an ENSEMBL ID}
\usage{
ensemblurl(ensemblid)
}
\arguments{
\item{ensemblid}{ENSEMBL ID for a gene}
}
\value{
A character referring to the ENSEMBL URL
}
\description{
It makes the url from ENSEMBL giving the ENSEMBL ID.
}
\seealso{
Other URL generation: 
\code{\link{dfGO}()},
\code{\link{entrezurl}()}
}
\concept{URL generation}
