% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriquepresa-URL-functions.R
\name{entrezurl}
\alias{entrezurl}
\title{ENTREZ URL generation
URL for an ENTREZ ID}
\usage{
entrezurl(entrezid)
}
\arguments{
\item{entrezid}{ENTREZ ID for a gene}
}
\value{
A character referring to the ENTREZ URL
}
\description{
It makes the url from ENTREZ giving the ENTREZ ID.
}
\seealso{
Other URL generation: 
\code{\link{dfGO}()},
\code{\link{ensemblurl}()}
}
\concept{URL generation}
