% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriquepresa-data.R
\docType{data}
\name{Eset50467}
\alias{Eset50467}
\title{Eset50467}
\format{
Large ExpressionSet
}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE50467}
}
\description{
Effect of SRSF2 over-expression in a human lung carcinoma cell line

Homo sapiens.

Expression Set from experiment GSE50467: Effect of SRSF2 over-expression in a human lung carcinoma cell line.
  SRSF2 is an important gene for the splicing of a great amount of genes.
Additionally, as it can bee seen in this study, the overexpression of this gene change the expression of a lot
  of genes.
The overall design includes 6 samples with SRSF2 over-expression and 6 samples as control of H358 cell line.
Expression data has been background-corrected and normalized.
In addition to expression data, phenoData, experimentData and featureData can also be found.
}
\examples{
data(Eset50467,package="enriquepresa")
}
\keyword{datasets}
