% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriquepresa-data.R
\docType{data}
\name{limma_df}
\alias{limma_df}
\title{limma_df}
\format{
data.frame
}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE50467}
}
\description{
Dataframe generated with the GSE50467 expression set.
This dataframe includes differential expression information for each probe.
The analysis was done with a t-test without equal variances assumed using limma package.
Dataframe includes the gene symbol, the t statistic, the LogFC, the p-value the adjusted p-value
  with both Benjamini-Hochberg and Bonferroni method, and the URL's to ENTREZ and ENSEMBL.
}
\examples{
data(limma_df,package="enriquepresa")
}
\keyword{datasets}
