% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriquepresa-plot-functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot generation}
\usage{
plot_volcano(adj_p, logFC, symbols, min_p = 0.05, diff = 1)
}
\arguments{
\item{adj_p}{information of adjusted p-values of a dataframe}

\item{logFC}{logFC for each gene of the same dataframe}

\item{symbols}{gene names in order to mark those that are differentially expressed}

\item{min_p}{adjusted p-value that is the limit of significance. By default is 0.05}

\item{diff}{limit of logFC to consider a gene interesting. By default is 1 (a gene must be two times
overexpressed or underexpressed)}
}
\value{
Volcano plot
}
\description{
It generates a Volcano plot for a set of genes. A volcano plot is a kind of representation where the genes
  that are differentially expressed over a p-value (here represented as the opposite decimal logarithm of the
  p-value) are remarked.
}
\concept{Volcano plot generation}
